+++ morse_source.c
@@ -6,6 +6,7 @@
 #include <string.h>
 #include <time.h>
 #include <math.h>
+#include <ctype.h>
 #ifdef __linux__
 #include "ws2811.h"
 #include <ctype.h>
@@ -48,7 +49,7 @@ static void MorseSource_convert_morse()
         char* m = morse_source.cmorse[mi];
         struct MorseChar* mc = &morse_source.morse[mi];
         mc->len = 0;
-        int ml = strlen(m);
+        size_t ml = strlen(m);
         for (int i = 0; i < ml; ++i)
         {
             if (m[i] == '-') {
@@ -140,7 +141,7 @@ void MorseSource_update_leds_morse(int frame, ws2811_t* ledstrip)
         }
         int letter_color = (index % 6) + 2; //<- 6 = number of letter colors, color 0 and 1 are reserved
         char* code = morse_source.cmorse[(int)c - 65];
-        int code_length = strlen(code);
+        size_t code_length = strlen(code);
         for (int ddi = 0; ddi < code_length; ++ddi)
         {
             char dd = code[ddi];
